Validation makes each row of the table have the same # of lines -> ARRAY
Do away with the string representation of the line centers following validation
becasue each row of the Table will now have the same size `xpos` element and
can be saved as a float directly.  This will help with the analysis because
there is not the need to "refloat" the string.

	modified:   flexure_analysis.py